Fixed some problems with the SoftTimer and other classes and finished implementing the SimpleTimer. Also added a new IntSetAction class which can be used to set Integer number options for menus.

+++ LCD.h
@@ -55,7 +55,7 @@ public:
    * @param digitRepresenter A 10 based number which reprepesents how many digits you 
    *    want to print(eg. 10 while print one zero + the number, 100 will print 00 + number and so on).
    */
-  void printFormatedNumber(int number, int digitRepresenter);
+  void printFormatedNumber(int number, long digitRepresenter);
   
   /**
    * Prints the number with a given number of chararacters infront (like P3 or XX5X).
@@ -64,7 +64,7 @@ public:
    *    want to print(eg. 10 while print one zero + the number, 100 will print 00 + number and so on).
    * @param character The character to be printed infront the number.
    */
-  void printFormatedNumber(int number, int digitRepresenter, int character);
+  void printFormatedNumber(int number, long digitRepresenter, int character);
   
   /**
    * Prints a string and moves the cursor back to the previous position.
